Move trigger definitions to shared-core.
