+++ packages/server/src/sdk/app/rowActions/crud.ts
@@ -9,7 +9,7 @@ import {
 } from "@budibase/types"
 import { generateRowActionsID } from "../../../db/utils"
 import automations from "../automations"
-import { definitions as TRIGGER_DEFINITIONS } from "../../../automations/triggerInfo"
+import { definitions as TRIGGER_DEFINITIONS } from "../../../../../shared-core/src/automations/triggers"
 import * as triggers from "../../../automations/triggers"
 import sdk from "../.."
