+++ packages/shared-core/src/automations/index.ts
@@ -1,1 +1,2 @@
 export * as steps from "./steps/index"
+export * as triggers from "./triggers/index"
