+++ packages/shared-core/src/automations/triggers/app.ts
