+++ packages/shared-core/src/automations/triggers/index.ts
