+++ packages/shared-core/src/automations/triggers/rowDeleted.ts
